+++ src/src/Infrastructure/Kentico/Xperience/Xperience/Services/ArticleService.cs
@@ -2,11 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using AutoMapper;
+using BizStream.Extensions.Kentico.Xperience.Retrievers.Abstractions.Documents;
 using BlogTemplate.Core.Abstractions.Models;
 using BlogTemplate.Infrastructure.Abstractions.Services;
 using BlogTemplate.Infrastructure.Kentico.Xperience.Abstractions;
 using BlogTemplate.Infrastructure.Kentico.Xperience.Abstractions.PageTypes;
-using BlogTemplate.Infrastructure.Kentico.Xperience.Abstractions.Retrievers;
 using BlogTemplate.Infrastructure.Kentico.Xperience.Extensions;
 using Microsoft.Extensions.Caching.Memory;
