+++ src/src/Mvc/Kentico/Xperience/Xperience/Extensions/IServiceCollectionExtensions.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Linq;
-using BlogTemplate.Infrastructure.Kentico.Xperience.Retrievers;
+using BizStream.Extensions.Kentico.Xperience.Retrievers.Documents;
 using BlogTemplate.Mvc.Kentico.Xperience.StaticWebAssetsStorage;
 using Kentico.Content.Web.Mvc;
 using Kentico.Content.Web.Mvc.Routing;
